chore: use gpt-4 as it is faster
It seems gpt-4 is faster than gpt-4-1106-preview for small inputs and
outputs. Might not be the case in terms of tokens per second, but gpt-4
seems to work better in this application.

Note that gpt-4 is more expensive, though.

gpt-3 family of models don't generate JSON output well or at all.

Another option is to consider function calling for JSON output with
gpt-3 in an assistant, but a first try didn't work well and is left as a
TODO. Note that getting JSON output is dependent on whether we use an
assistant or not. Needs further investigation. In any case, gpt-4 family
of models usually do the right thing and output a JSON.